Handle relations with _ (fixes #14)
relate() was assuming that all relations were ModuleName.lower(), however in at
least one case this is not true. Accounts has a relation to EmailAddresses
called 'email_addresses'. This probes the relationships dictionary to see if
.lower() matches, if not it tries the transformation from ModuleName =>
module_name and sees if it is in there, if not it just returns ''.